bump versions for site, upgrade API usage, move html to avoid conflicts
