+++ website/src/theme/BlogPostItem/Header/Authors/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import clsx from 'clsx';
-import {useBlogPost} from '@docusaurus/theme-common/internal';
+import {useBlogPost} from '@docusaurus/plugin-content-blog/client';
 import BlogPostItemHeaderAuthor from '@theme/BlogPostItem/Header/Author';
 import styles from './styles.module.css';
 // Component responsible for the authors layout
