+++ website/src/theme/BlogPostItem/Header/Info/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import clsx from 'clsx';
-import {useBlogPost} from '@docusaurus/theme-common/internal';
+import {useBlogPost} from '@docusaurus/plugin-content-blog/client';
 import styles from './styles.module.css';
 
 function Date({date, formattedDate}) {
