+++ website/src/theme/BlogPostItem/Header/Title/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import clsx from 'clsx';
 import Link from '@docusaurus/Link';
-import {useBlogPost} from '@docusaurus/theme-common/internal';
+import {useBlogPost} from '@docusaurus/plugin-content-blog/client';
 import styles from './styles.module.css';
 export default function BlogPostItemHeaderTitle({className}) {
   const {metadata, isBlogPostPage} = useBlogPost();
